Refactor: 배치 파일을 잡 목적에 따라 분리
